Use editorAreaResize handler instead of manipulating editor inside PanelManager
